requirements.txt
python-dotenv
google-generativeai
langgraph
streamlit
mysql-connector-python
pymysql